pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "schw_api_service"
version = "0.1.0"
description = "Lightweight Charles Schwab API client for auth, accounts, and market data"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "SwapLaunch" }
]
dependencies = [
  "requests>=2.31",
  "python-dotenv>=1.0",
  "cryptography>=41",
  "filelock>=3.12",
  "loguru>=0.7"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]


